fix: close the remaining MSVC /W4 warnings in the resampler and WAV loader
- ResamplingContainer: pass float rates to the LanczosResampler constructor
  explicitly (it takes float; the double to float narrowing was implicit
  inside std::make_unique).
- wav: WAVEFORMATEXTENSIBLE::dwChannelMask is a 32-bit DWORD. The field was
  declared uint16_t, so the value read from the file was truncated to 16
  bits. Nothing reads it yet, so no behaviour changes.

+++ dsp/ResamplingContainer/ResamplingContainer.h
@@ -141,8 +141,9 @@ public:
     }
 
     {
-      mResampler1 = std::make_unique<LanczosResampler>(mInputSampleRate, mRenderingSampleRate);
-      mResampler2 = std::make_unique<LanczosResampler>(mRenderingSampleRate, mInputSampleRate);
+      // LanczosResampler takes float rates; the narrowing is intended and now explicit
+      mResampler1 = std::make_unique<LanczosResampler>(static_cast<float>(mInputSampleRate), static_cast<float>(mRenderingSampleRate));
+      mResampler2 = std::make_unique<LanczosResampler>(static_cast<float>(mRenderingSampleRate), static_cast<float>(mInputSampleRate));
 
       // Zeroes the scratch pointers so that we warm up with silence.
       ClearBuffers();
